requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.2
pandas==2.1.3
scikit-learn==1.3.2
nltk==3.8.1
google-generativeai==0.3.2